pyproject.toml
[tool.poetry]
name = "trello-reporter"
version = "0.1.0"
description = "Send email reports about old Trello cards"
authors = ["Alex Hamlin <alex@alexhamlin.co>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
boto3 = "^1.9.151"
python-dateutil = "^2.8.0"
requests = "^2.22.0"

[tool.poetry.dev-dependencies]
black = "=19.3b0"
pylint = "^2.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
